refactor(core): remove legacy `ProbInput` class and related tests
Remove the obsolete `ProbInput` implementation
and its associated test cases,
replacing it with a modernized implementation structure.
Updates all dependent modules to use the revised `ProbInput`.

Closes: #536
Ref: #533
