+++ src/uqtestfuns/core/registry/resolver.py
@@ -22,7 +22,7 @@ from .specs import (
 from uqtestfuns.core.prob_input.marginal import Marginal
 from uqtestfuns.core.parameters import Parameters
 from uqtestfuns.core.registry.parser import SpecValidationError
-from uqtestfuns.core.prob_input.probabilistic_input_new import ProbInput
+from uqtestfuns.core.prob_input.probabilistic_input import ProbInput
 from uqtestfuns.core.registry.parser.utils import substitute_idx
